Move PermissionParserException to Bolt\Exception
