+++ src/PermissionParser.php
@@ -3,7 +3,7 @@
 namespace Bolt;
 
 use Bolt\Exception\PermissionLexerException;
-use Bolt\PermissionParserException;
+use Bolt\Exception\PermissionParserException;
 
 /**
  * Lexer and parser for permission query syntax.
